fix(scene-graph): harden clone prop copies
