+++ tests/engine/scene-graph/clone-binding-corruption.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from 'bun:test'
 
-import { SceneGraph, type SceneNode } from '@open-pencil/core'
+import { SceneGraph } from '@open-pencil/core'
 import { collectSubtrees } from '@open-pencil/core/editor/clipboard/subtree-history'
 
 function pageId(graph: SceneGraph): string {
@@ -390,189 +390,5 @@ describe('variable deletion and re-clone safety', () => {
     const clone2 = graph.cloneTree(node.id, pageId(graph))
     expect(clone2).toBeDefined()
     expect(graph.getNode(clone2.id).boundVariables['fills/0/color']).toBeUndefined()
-  })
-})
-
-// ─── cloneNodeProps deep-copy verification ──────────────────────────────────
-
-describe('cloneNodeProps deep-copies vectorNetwork', () => {
-  test('cloned vectorNetwork is not the same reference as original', () => {
-    const graph = new SceneGraph()
-    const node = graph.createNode('RECTANGLE', pageId(graph), {
-      name: 'Vector',
-      vectorNetwork: {
-        vertices: [
-          { x: 0, y: 0 },
-          { x: 100, y: 0 }
-        ],
-        segments: [{ start: 0, end: 1, tangentStart: { x: 0, y: 0 }, tangentEnd: { x: 0, y: 0 } }],
-        regions: [{ windingRule: 'NONZERO' as const, loops: [[0, 1]] }]
-      }
-    })
-    const original = graph.getNode(node.id)
-    const clone = graph.cloneTree(node.id, pageId(graph))
-    const clonedNode = graph.getNode(clone.id)
-
-    expect(clonedNode.vectorNetwork).not.toBe(original.vectorNetwork)
-    if (original.vectorNetwork && clonedNode.vectorNetwork) {
-      expect(clonedNode.vectorNetwork.vertices).not.toBe(original.vectorNetwork.vertices)
-      expect(clonedNode.vectorNetwork.segments).not.toBe(original.vectorNetwork.segments)
-      expect(clonedNode.vectorNetwork.regions).not.toBe(original.vectorNetwork.regions)
-    }
-  })
-
-  test('mutating clone vectorNetwork does not affect original', () => {
-    const graph = new SceneGraph()
-    const node = graph.createNode('RECTANGLE', pageId(graph), {
-      name: 'Vector',
-      vectorNetwork: {
-        vertices: [
-          { x: 0, y: 0 },
-          { x: 100, y: 50 }
-        ],
-        segments: [{ start: 0, end: 1, tangentStart: { x: 0, y: 0 }, tangentEnd: { x: 0, y: 0 } }],
-        regions: []
-      }
-    })
-    const original = graph.getNode(node.id)
-    const clone = graph.cloneTree(node.id, pageId(graph))
-    const clonedNode = graph.getNode(clone.id)
-
-    // Mutate clone's vertex
-    if (clonedNode.vectorNetwork) {
-      clonedNode.vectorNetwork.vertices[0].x = 999
-    }
-    // Original must be unaffected
-    expect(original.vectorNetwork?.vertices[0].x).toBe(0)
-  })
-})
-
-describe('cloneNodeProps deep-copies textPicture and figmaDerivedTextGlyphs', () => {
-  test('cloned textPicture is not the same reference as original', () => {
-    const graph = new SceneGraph()
-    const node = graph.createNode('TEXT', pageId(graph), {
-      name: 'Text',
-      text: 'Hello',
-      fontSize: 14,
-      fontFamily: 'Inter',
-      textPicture: new Uint8Array([1, 2, 3, 4])
-    })
-    const original = graph.getNode(node.id)
-    const clone = graph.cloneTree(node.id, pageId(graph))
-    const clonedNode = graph.getNode(clone.id)
-
-    expect(clonedNode.textPicture).not.toBe(original.textPicture)
-    if (original.textPicture && clonedNode.textPicture) {
-      // Mutate clone's data
-      clonedNode.textPicture[0] = 99
-      // Original must be unaffected
-      expect(original.textPicture[0]).toBe(1)
-    }
-  })
-
-  test('cloned figmaDerivedTextGlyphs is not the same reference as original', () => {
-    const graph = new SceneGraph()
-    const node = graph.createNode('TEXT', pageId(graph), {
-      name: 'Text',
-      text: 'Hi',
-      fontSize: 14,
-      fontFamily: 'Inter',
-      figmaDerivedTextGlyphs: [
-        { commandsBlob: new Uint8Array([10, 20]), x: 0, y: 0, fontSize: 14 },
-        { commandsBlob: new Uint8Array([30, 40]), x: 10, y: 0, fontSize: 14 }
-      ]
-    })
-    const original = graph.getNode(node.id)
-    const clone = graph.cloneTree(node.id, pageId(graph))
-    const clonedNode = graph.getNode(clone.id)
-
-    expect(clonedNode.figmaDerivedTextGlyphs).not.toBe(original.figmaDerivedTextGlyphs)
-    if (original.figmaDerivedTextGlyphs && clonedNode.figmaDerivedTextGlyphs) {
-      expect(clonedNode.figmaDerivedTextGlyphs[0]).not.toBe(original.figmaDerivedTextGlyphs[0])
-      // commandsBlob should be independent
-      expect(clonedNode.figmaDerivedTextGlyphs[0].commandsBlob).not.toBe(
-        original.figmaDerivedTextGlyphs[0].commandsBlob
-      )
-      // Mutate clone's blob
-      clonedNode.figmaDerivedTextGlyphs[0].commandsBlob[0] = 99
-      // Original must be unaffected
-      expect(original.figmaDerivedTextGlyphs[0].commandsBlob[0]).toBe(10)
-    }
-  })
-})
-
-// ─── cloneNodeProps deep-copies overrides values ─────────────────────────────
-
-describe('cloneNodeProps deep-copies overrides values', () => {
-  test('mutating clone overrides does not affect original', () => {
-    const graph = new SceneGraph()
-    const page = pageId(graph)
-    // Create component with child that has fills
-    const component = graph.createNode('COMPONENT', page, { name: 'Comp', width: 100, height: 40 })
-    const _child = graph.createNode('RECTANGLE', component.id, {
-      name: 'Bg',
-      width: 100,
-      height: 40,
-      fills: [{ type: 'SOLID', color: { r: 1, g: 0, b: 0, a: 1 }, visible: true, opacity: 1 }]
-    })
-
-    // Create instance
-    const instance = graph.createInstance(component.id, page)
-    if (!instance) throw new Error('instance failed')
-    const instanceChild = graph.getChildren(instance.id)[0]
-
-    // Set an override with an object value (fills array)
-    instance.overrides[`${instanceChild.id}:fills`] = [
-      { type: 'SOLID', color: { r: 0, g: 0, b: 1, a: 1 }, visible: true, opacity: 1 }
-    ]
-
-    // Clone the instance
-    const clone = graph.cloneTree(instance.id, page)
-    if (!clone) throw new Error('clone failed')
-    const clonedInstance = graph.getNode(clone.id)
-
-    // Mutate the clone's overrides value
-    const overrideKey = `${instanceChild.id}:fills`
-    const cloneOverrideVal = clonedInstance.overrides[overrideKey] as Array<{
-      color: { r: number }
-    }>
-    if (cloneOverrideVal) {
-      cloneOverrideVal[0].color.r = 0.5
-    }
-
-    // Original instance overrides must be unaffected
-    const origOverrideVal = instance.overrides[overrideKey] as Array<{ color: { r: number } }>
-    if (origOverrideVal) {
-      expect(origOverrideVal[0].color.r).toBe(0) // still blue, not mutated
-    }
-  })
-})
-
-describe('cloneNodeProps handles undefined arrays defensively', () => {
-  test('cloneNodeProps does not crash when array fields are undefined', () => {
-    const graph = new SceneGraph()
-    const page = pageId(graph)
-    const node = graph.createNode('RECTANGLE', page, { name: 'Rect' })
-    const raw = graph.getNode(node.id)
-    // Simulate a node from .fig parsing where array fields could be undefined
-    const patch: Partial<
-      Pick<
-        SceneNode,
-        | 'componentPropertyDefinitions'
-        | 'symbolLinks'
-        | 'exportSettings'
-        | 'pluginData'
-        | 'pluginRelaunchData'
-      >
-    > = {
-      componentPropertyDefinitions: undefined,
-      symbolLinks: undefined,
-      exportSettings: undefined,
-      pluginData: undefined,
-      pluginRelaunchData: undefined
-    }
-    Object.assign(raw, patch)
-    // cloneTree must not throw TypeError on .map() of undefined
-    expect(() => graph.cloneTree(node.id, page)).not.toThrow()
   })
 })
